Refactored some file structuring
