+++ frontend/src/App.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Outlet } from "react-router";
-import Nav from "./ui/Nav";
+import Nav from "./components/Nav";
 
 function App() {
