+++ frontend/src/css/home_styles.css
@@ -6,23 +6,38 @@ body, html {
 }
 
 .PatientsPage {
+    position: relative;
     background-image: url('../img/nurses-stock.jpg');
     background-size: cover;
     width: 100%;
     height: 100%;
     padding: 0;
     margin: 0;
+    z-index: 1;
 }
 
 .PatientsPage::before {
     content: "";
     position: absolute;
+    top: 0;
+    left: 0;
     width: 100%;
-    height: 120%;
+    height: 100%;
     background-color: rgba(0, 0, 0, 0.8);
+    z-index: 1;
 }
 
 .header {
+    position: relative;
     background-color: black;
+    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+    font-size: 64px;
     color: white;
+    z-index: 2; 
+    padding: 20px; 
+}
+
+.container-fluid {
+    position:relative;
+    z-index: 3; /*this has to be a higher value than its child components */
 }
