+++ frontend/src/main.jsx
@@ -2,7 +2,6 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx';
-import Nav from './ui/Nav.jsx';
 import Patients from './routes/Patients.jsx'
 import Login from './routes/Login.jsx'
 import WeatherForecast from './routes/WeatherForecast-test.jsx';
