+++ frontend/src/routes/Patients.jsx
@@ -1,11 +1,30 @@
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import PatientCard from '../ui/PatientCard.jsx';
+import PatientCard from '../components/PatientCard.jsx';
 import '../css/home_styles.css';
 import axios from 'axios';
 
 
 const Patients = () => {
+
+// const [patientData, setPatientData] = useState([]);
+
+    // useEffect(() => {
+    //     const fetchData = async () => {
+    //         try {
+    //             console.log('Fetching Patient data...');
+    //             // const response = await axios.get('http://localhost:5232/');
+    //             console.log('Response:', response);
+    //             setPatientData(response.data);
+    //         } catch (error) {
+    //             console.error('Error fetching data:', error);
+    //             console.error('Axios error config:', error.config);
+    //             console.error('Axios error request:', error.request);
+    //             console.error('Axios error response:', error.response);
+    //         }
+    //     };
+    // }, []);
+
 
 
     return(
@@ -43,6 +62,5 @@ const Patients = () => {
         </>
     )
 }
-
 export default Patients;
